Remove fwupd from core-deps
 (a) It's broken:

    Running custom install script '/buildstream/build/po/make-images.sh'
    Failed to run install script '/buildstream/build/po/make-images.sh'
    FAILED: meson-install

 (b) I don't know why

 (c) I don't want to fix it

 (d) We have a history of problems with this module

 (e) firmware updates are not likely to work very well within the
BuildStream sandbox

So I'm just going to remove this.

It would be *desirable* to bring it back, if somebody who cares to fix
it wants to do so.  Feel free to bring it back if you can get the build
working in both git master and the latest tarball release.