Use -Ddisable_gtkdoc=true just for building GStreamer
I swear I made up this flag when adding it to JHBuild's default
configuration, but apparently GStreamer was actually using it. It
doesn't match our preferred style for meson arguments, though, so we
don't want to pass this to all modules... use it just for GStreamer.

This is particularly important because the gtk-doc build for
gst-plugins-base is currently broken.