project.conf: Provide access to pulseaudio from within `bst shell`
This leverages some new features to expose the host XDG_RUNTIME_DIR
into the runtime where the pulseaudio socket will reside, and to
set the PULSE_SERVER environment so that applications using
pulseaudio will be able to talk to the host side audio server.

Also, with format version 4 we remove `environment-inherit` as it
is redundant with the newly added `environment` configuration.