Reorganize core metamodules
Remove meta-gnome-core, meta-gnome-devel-platform, and
meta-gnome-extended-devel-platform.

core.bst now includes only meta-gnome-core-shell.bst,
meta-gnome-core-utilities.bst, and meta-gnome-core-os-services.bst. This
way, we guarantee that nobody will forget to update the metamodules
anymore, as was a common problem with JHBuild.

I'm not sure if having a separate meta-gnome-core-os-services.bst is
actually useful or not, but let's leave it for now.

Fixes #4