Upgrade freedesktop-sdk and switch to sidebranch
We require the ability to make immediate fixes to the freedesktop SDK,
so following the master branch is not currently suitable for us. Use a
sidebranch instead. It's current git master, plus one commit to upgrade
GCC to make WebKit trunk buildable.

In the future, we can reconsider whether having our own branch is
needed, but I expect this is going to be required at least on release
days as the release manager needs the ability to make immediate changes
throughout the entirety of the stack.

This additionally fixes #35, so that our CI will have a chance to
succeed once again, and hopefully fixes #30.