Host popt tarball on people.gnome.org
rpm5.org has been down all weekend, we can no longer use it as upstream.
Please never add this back, it's too unreliable.

There seem to be no suitable websites hosting this tarball, so we'll
have to do so ourselves. Ideally we wouldn't need to do this under my
personal namespace, and would find a better place to host problematic
tarballs in the future, but this will do for now.