sdk/WebKitGTK+.bst: Remove max-jobs setting
Since recent versions of BuildStream, the max-jobs is automatically
capped at 8, so it will not blow out of proportion on build servers
with 96 cores and the like.

Also, max-jobs is an internal protected variable in BuildStream,
an upcoming update intends to cause setting of this internal detail
to raise an error.