core-deps/boost.bst: Fix max-jobs usage to avoid random cache key
The `max-jobs` variable is an automatic variable which should
not be fed to builds directly in the build commands, this will
cause random rebuilds to occur depending on how many jobs BuildStream
has decided to use, based on the host.

Fixed this by using the `environment-nocache` feature which was
designed specifically for this.