sdk/cogl.bst: do not build the gstreamer plugin
This plugin is bitrotten and doesn't function well. Sadly it
also seem to have a waaaay to high priority rank for no damn
reason. We can, *and should* do away without it and probably
nobody will miss it. Distributions like fedora, as of f30,
don't seem to build it at all.

Close:
https://gitlab.gnome.org/World/podcasts/issues/132
https://gitlab.gnome.org/GNOME/gnome-build-meta/issues/82