core/gdm.bst: Added custom GDM pam configuration
Our PAM configuration resembles the redhat one, but we do not
use selinux, and we do not build PAM with the third party patch
which enables the pam_console.so module, so we need a different
variant.

Eventually we could upstream a GNOME configuration for
GDM integration into GNOME system images.