.gitlab-ci.yml: Set `--max-jobs` on the command line.
Use a larger number of cores when running on workers with many cores.

Instead of setting the BST variable to include --max-jobs, we only
set `--max-jobs` in the line where we actually run `bst build`. This
is just a workaround for a technical difficulty with gitlab variables,
trying to update the BST definition just results in $(( $(nproc) / 4 ))
being given to `bst` literally without any shell expansion.

This fixes #165