Add ABI checker from freedesktop-sdk
Adds the ABI checker as it currently stands in freedesktop-sdk. An ABI
check is also added to the CI in the `reports` stage. In order for this
to actually be run the variable $STABLE_ABI needs to be set, as we don't
want this to fail on master.

The ABI check is run on `abi/abi-image.bst`, which is an image
constructed from the platform and the headers of the SDK. This is how
things are done in freedesktop-sdk also.

There is also `abidiff-suppressions.ini`, which allows for configuration
of the ABI checker to ignore some libraries or symbols. This is copied
from freedesktop-sdk also, as otherwise we would get failures from
breakages allowed in freedesktop-sdk.