core-deps: Add tracker-3 and tracker-miners-3 elements
We are targetting GNOME 3.38 to include Tracker 3.0, and we want
Tracker 3.0 available in the master Flatpak runtimes now so apps
can start porting and testing.

Tracker 2 and 3 are parallel installable, however we don't want
to depend on on both versions in a stable GNOME release. Our plan
is to port all apps in core to use Tracker 3 and to remove Tracker 2.0
before the GNOME 3.37.90 release. (Tracking issue here:
https://gitlab.gnome.org/GNOME/tracker/-/issues/194)