vm: Increase the size of the boot partition
For the moment the kernel with initramfs is around 35M. If there are
3 successive updates that include kernel/initramfs updates, then
for a short time the /boot disk will need to have 3 sets of kernels.
This is more than 100M so it fails.

Also it is likely that the kernel and initramfs will slowly increase
with the time. Resizing a boot partition is a bit too annoying for
most people to do. So 500M should be a safe bet for a long time,
rather than going for 200M.