sdk/{gjs,glib,gtk+-3,gtk,libsoup, libsoup-2.4, mutter}.bst: build with sysprof support
This commit also moves sysprof-minimal from sdk-deps to sdk.bst due to the
sysprof-capture dependency ending up in Requires.private .pc files of modules
that link against it, even when they statically link the library.

This makes it so that pkg-config refuses to use the .pc files of
the libraries unless sysprof-capture is also available, and thus
forces it to be part of the Sdk now.

More details can be found here:

https://github.com/mesonbuild/meson/issues/3970

Fixes #294

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/770>