WIP: Use Tracker 3 by default, Tracker 2.3 for GNOME Photos
For GNOME 3.38, we expect to have all core apps ported to Tracker 3
except for GNOME Photos. The Photos port to Tracker 3 lacks review
and testing so for reliability's sake we should stick to 2.x.

We must avoid Tracker 2 and 3 indexing the same locations as much
as possible. To this end, we will ship Tracker 2.x with autostart
disabled, and GNOME Photos will start Tracker miners on-demand.
GNOME Online Miners will also continue to work against Tracker 2.x.