core-deps/qemu.bst: Enable SDL and GTK
virtio-gpu did not seem to work with Boxes (using spice). In order
to debug that I tried to enable GTK with GL acceleration as well
as SDL to compare. But that made virtio-gpu work with spice.
It is possible that QEMU does not enable virtio-gpu when only spice
is configured.

Having SDL and GTK backends can be useful anyway when calling QEMU
directly since it is a bit simpler in the command line.