ci: add a manual s3 job
The idea is that we want a static url that will provide a working
image. The job is manual as for the time being we want someone
to manually verify that its working so we don't give people
broken images.

Hopefully in the future we can have an automated testsuite
for this like openQA.