gnome-shell: add a private version of newer pulseaudio
3.36.7 bumped the required version of pulseaudio,
from the gvc submodule in order to avoid a deadlock under
Xwayland.

The 19.08 version of fd.o-sdk sadly only inclues v12.2 of pulseaudio,
so we need to bring in a private copy so we can built the shell, as
its not worth reverting the gvc commit as pulse's behavior is
broken.

The pulse elements added are adapted copies from the ones in
fd.o-sdk 20.08. We could have got away with a single element
but its simpler imo to copy-paste. There are only a couple months
left in the supported lifepsan of 3.36 branch.

Thankfully we don't produce a vm image for 3.36, as it would
have been half broken linking half things against pulse v12
and shell against v13. So we can get away with just having
succesful builds. Downstreams are encouraged to upgrade
to pulse 13 or later.

See:

https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/3250

https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/issues/693