project.conf: Don't inherit rarely set env-vars in shell's toolbox mode
When one of those environment variables wasn't set at all on the host,
it would be set i.e. for DBUS_SYSTEM_BUS_ADDRESS, its value will be
$DBUS_SYSTEM_BUS_ADDRESS in the buildstream shell, instead of being left
unset (where the default as per the DBus spec would be choosen instead
if left unset).

So this commit gathers potentially problematic environment variables
together, comments them and add a comment for people reading this so
it can still be inherited if needed in this set-up.