project.conf: disable bmalloc and JIT during build
We recently upgraded our aarch64 CI runner, and now it is running
CentOS. This is a problem because RHEL and CentOS on aarch64 use a 64 KB
page size, unlike all other distros, which use 4 KB. The mismatch in
expected page size between the container and the host is causing
problems for gobject-introspection. gi-scanner runs during the build and
calls every get_type() function, which winds up initializing WebKit and
bmalloc. But bmalloc requires the system page size is 16 KB or less.

The normal solution for this is to use the -DUSE_64KB_PAGE_BLOCK=ON
option, which exists specifically to make WebKit not crash on aarch64
CentOS and RHEL. It shouldn't be used elsewhere because (a) 64 KB page
blocks are 16x bigger than the usual 4 KB page blocks, which is bad for
performance and memory usage on embedded devices, and (b) disabling
bmalloc disables all of WebKit's heap security features. So we only want
to do that as an absolute last resort.

Fortunately, we only need to disable bmalloc during the build itself.
The WebKitGTK that we've built will work just fine as long as it's not
itself run on CentOS or RHEL.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1170>