Add CI stage to test x86 iso images in openqa
This is expected to run on 'master' branch as an optional, manual job.
Please use it on any MR that might trigger a regression in the VM
images, and wherever else it might be useful.

Documentation at https://gitlab.gnome.org/GNOME/gnome-build-meta/-/wikis/openqa/OpenQA-for-GNOME-developers

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1251>