Update Aliases: reorganize aliases in project.conf
Separate out the git aliases from the 'file' aliases in project.conf.
This helps ensure that no one alias is used for both git sources, and
file-download sources (eg tar, remote, zip sources). This is important,
because we will need to use different kinds of urls as the mirror
aliases, when mirrors are set up.