vm/filesystem-devel: include debuginfo by default
Include debuginfo for most of the stack, minus webkit.

Also strip the generated sources to reduce the size of the
refs/objects.

And also strip the .debug files of various utilities that
are sizable enough and we are unlikely to debug in gnomeos
as they are outside our stack.

Close #322

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1499>