EOL the 3.38 runtime, take two
The original EOL was reverted in a9c6bc60 and never reinstated, but we
stopped updating the runtime anyway because of course we did: there were
no more scheduled updates! EOL it once again using the date of the final
update to the runtime.

The four month gap with no updates before we remembered to add the EOL
notice is unfortunate, but can't do anything about that now.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1575>