core/gnome-maps.bst: Remove dependencies for GFBGraph and g-o-a
GFBGraph hasn't been used by Maps since 41 when we dropped
support for Facebook check-in (logging in in g-o-a had been
broken for a while).
And for 43, we have removed support for Foursquare check-ins
(and thus the dependency on g-o-a also got removed).
Now, explicitly depend on WebkitGTK (as we're still
depending on it for the OSM editing login).

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1585>