core/gnome-initial-setup.bst: switch to libsoup 3 dependencies
gnome-initial-setup is now able to use webkit2gtk-4.1 (the libsoup 3,
GTK 3 API version of WebKitGTK) and librest-1.0 (which optionally
supports libsoup 3, and which we build with libsoup 3 enabled).

This is needed to avoid crashing on startup, since gnome-initial-setup
links to geocode-glib, which we have already switched to libsoup 3.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1631>