world: purge elements that depend on webkit2gtk-4.0
These elements still depend on webkit2gtk-4.0, which we want to remove
so that we only need to build WebKitGTK twice instead of thrice as we
are currently doing. Since they are in world and are outside the purview
of release team, let's just remove them. Feel free to add them back in
the future if they are ported to webkit2gtk-4.1, except:

 * bijiben is now gnome-notes and should be named accordingly if
   restored
 * gnome-documents is archived and should probably not be restored

This commit additionally removes snowball, which was only used by geary,
and gnome-online-miners, which was only used by gnome-documents