ci: Add EXTRA_TESTS custom pipeline variable
This variable can be set in scheduled pipelines and when manually
starting pipelines. When set to `openqa` it causes the
iso-installer-x86_64 and test-iso-installer-x86_64 jobs to run
by default, providing a one-step method to trigger an OpenQA test
cycle.

This is a big improvement on the current method which can take
several hours and involves: (1) waiting for 'track' and 'build' to
complete, (2) starting 'iso-installer-x86_64' manually, (3) waiting for
that to complete, (4) finally, starting 'test-iso-installer-x86_64'.

Fixes https://gitlab.gnome.org/GNOME/gnome-build-meta/-/issues/430