Move all machine config into `POST isos/` call
Setting machine config in the call to `POST machines/` is not useful.
The config is set up correctly on the new machine. But the tests
run against the machine `qemu_x86_64` that's predefined on
openqa.gnome.org, and we can't override that. (Setting `MACHINE` when
starting the job makes the job fail to start).

We should have config in Git as much as possible, so, I have removed
config from the predefined `qemu_x86_64` machine and we now set it
in the `start_job.sh` script that calls `POST isos/`.

See: http://open.qa/docs/#_variable_precedence