Update to WebKitGTK 2.39.1 and webkitgtk-6.0 API version
The webkitgtk-6.0 API version is just a new name for webkit2gtk-5.0. It
fulfills a longstanding goal of removing the internal 2, and will
hopefully avoid confusion caused by applications that have begun to
depend on webkit2gtk-5.0 before the API was stable. The goal is for
webkitgtk-6.0 to be actually stable by the time it lands in the upcoming
stable WebKitGTK 2.40.0.

This commit adds unifdef.bst, a new build dependency, and libavif.bst, a
new runtime dependency. The libavif.bst is based on Phil's from the WebKit
SDK.

I've stopped disabling evolution's GTK 4 web view support with this
commit.

Finally, note that I've disabled the AVIF image support because we don't
have libavif yet. We need to package that up.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1864>