Revert "elements/freedesktop-sdk.bst: Use freedesktop-sdk master"
This reverts commit 40a6f53e008ee1b60bd0d078d2f455adc9374ec9.

We don't have a working build pushed into flathub-beta so we
can't build the Sdk.Extensions the Applications need in CI, but
at the same time the runtime was updated so nightly applications
can't run atm, nor rebuild and push on the CI to fix it.

Revert till Monday when people are around.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2137>