Remove START_AFTER_TEST
This is causing the gnome_apps test to be skipped if gnome_install
fails, which is not what we want.

The idea was to ensure tests run in a specific order, but it doesn't
actually matter what order they run in. As long as WORKER_CLASS is set
then they will all still run sequentially on the correct runner.