flatpak: Absorb the documentation into the .Sdk
When shipping the Sdk, we want to provide the Documentation
by default so Tools/IDEs can consume it as is. The Sdk should be
coming with batteries included rather than a lego set.

Currently, applications like Workbench have to resort into
copying all the documentation at build time, it's not guaranteed
the .Sdk.Docs extension will exist at runtime.

Close #710