sdk/libjxl: disable gdk-pixbuf loader for gnome 45
It's not much tested and it's causing a couple crashes. In addition
we were backporting patches on top of 0.8 release in order to build
the pixbuf loader, but that's not sustainable for the stable runtime
build.

We can try enabling it again once there's a 0.9 release and we branch
out gnome 45.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2296>