Enroll Microsoft keys by default
To make it safer to install on systems with buggy firmware signed but
Microsoft keys instead of firmware keys, we need the Microsoft keys
enrolled. However, we keep a set of keys without Microsoft keys
for users that do not trust them.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2220>