Update the GitHub repository hosting our kernel
We need to temporarily update the repository hosting the megous kernel.
This is because the original repository is being closed down.

We have the intention to mirror the original repository into GNOME's
internal lorry.

However, we have several issues related to the PinePhone Pro & PinePhone.
(See #725 for a summary.)

We will drop this commit once we confirm a working build, and use
the lorry mirroring service.