.gitlab-ci/buildstream.conf: don't cache sources in the dedicated volume
The git_repo plugin is pretty efficient at downloading just what it needs
when fetching (most of the time), and is quite bad at recovering when its
cache is "corrupted".

We can also enable source cache, as the git_repo plugin is usable with the
source cache.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2255>