Remove the libphonenumber-build.patch file
We previously used a patch file for libphonenumber because we needed
the patch file for <v8.10.16>.

The patch file was provided by Julian Sparber to remove the Java
dependencies when building the cpp edition of libphonenumber.

In the next commit, we will update the tag for libphonenumber to a more
recent version, i.e., 2023 versus the original 2019 tag.

The new version provides a direct CMake flag: <DREGENERATE_METADATA>,
which will achieve the same outcome as the patch file.