Add additional CMake arguments to libphonenumber
We need the following CMake arguments inside libphonenumber:
- USE_BOOST
- BUILD_TESTING
- FETCHCONTENT_FULLY_DISCONNECTED

These CMake arguements are required because there is a linking issue
when using the new version of protobuf, which fails to link abseil
when building.

- https://github.com/google/libphonenumber/pull/3215

Above is a link that will take you to an upstream pull request that
is yet to be completed.

In the meantime, we can still achieve a successfull build by following:
- https://github.com/flathub/org.gnome.Calls/blob/master/org.gnome.Calls.yaml