Add a patch file for protobuf-base to fix libphonenumber
As mentioned previously, we are following the flathub/org.gnome.Calls
setup. This involves editing protobuf-base from the freedesktop-sdk
junction to include a specific set of CMake arguments.

These arguments, and alongside the ones from the
world/libphonenumber.bst are required because protobuf automatically
installs abseil, and the build fails due to linking issues.