Introduce a patch file for libphonenumber builds
We need a patch file applied to the freedesktop-sdk junction for:
- elements/components/abseil-cpp.bst
- elements/components/protobuf-base.bst

The patch file is needed because libphonenumber is failing to compile
when linking the abseil library to protobuf.

This takes inspiration from the flathub/org.gnome.Calls.

Please note that this patch file will be removed once we update the
freedesktop-sdk junction in an upstream MR, and update the tracking
ref for the junction inside this repository. #746 tracks this as a
follow-up task, once this MR is accepted.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2409>