Ignore the venv folder recursively
Given that GNOME OS is now supported by Buildstream 2.0, and that
Buildstream 2.0 is now supported by the Pip package manager:

It would be useful if we could ensure that the venv folder is ignored
by Git. The reasoning is because Git will begin tracking the venv. This
is problematic when running the <git add -A> command.

The folder naming is dependent on developers. However, the standard
across Python seems to be the keyword <venv>. We obviously cannot
support all possible variations.

However, at the bare minimum, we can support ignoring the industry
standard for naming virtual environments, i.e., <venv>.