Remove the megous alias
The megous alias is useless to the GNOME OS repository because the
contents, i.e., the linux-firmware and kernel files are essentially
being removed from the URL.

As of 2024 onwards, we cannot rely on this domain to provide us with
the necessary files to build and run the PinePhone Pro & PinePhone boards.

We have: #748, as the main goal to research, and work out a mechanism
for integrating the freedesktop-sdk kernel instead. However, this is not
a straight-forward procedure.

Thus, we will replace the linux-firmware, and kernel with forks. This will
be done in the next commit, and should be noted that they are temporary
solutions, whilst we figure out a long term solution inside #748.