Update the forks related to the megous kernel files
As mentioned in the previous commit, we intend to temporarily update the
necessary repositories that are hosting our required kernel and firmware.

This is a temporary measure because the original megous URL is now
terminated.

Our long-term solution is to integrate the freedesktop-sdk repositories
kernel for our PinePhone Pro and PinePhone boards. Please see #748 for
additional information on this.

- There is external GitHub fork exists for the kernel
- There is an external GitLab fork for our Linux Firmware files
  - This GitLab fork is maintained by us

Naturally, this is a temporary solution to get the builds working.