freedesktop-sdk.bst: Add PinePhone Pro config patch
This is a temporary patch file that will be used to activate the
required kernel configs in order to boot the PinePhone Pro and
PinePhone boards.

This is a link to an upstream backport (release/23.08) for these configs:
- https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/merge_requests/18908

We will drop this patch file once we track the configs through our
junction.

Additionally, we will also update the kernel config tests in the next
commit for the PinePhone Pro to reflect the patch file, and eventually,
the upstream MR.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2455>