utils/test_phone_image.py: Refine PinePhone Pro configs
We are editing the tests for the PinePhone Pro to reflect the changes
made during the work from #755.

The work from #755 was based on swapping the megous kernel for the
freedesktop-sdk kernel, and integrating a single kernel for both mobile
boards inside GNOME.

This commit simply updates the kernel configs that are required to be
inside the PinePhone Pro image.

The following is a link to an upstream MR where these configs have been
merged in and will eventually be tracked by GNOME OS:
- https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/merge_requests/18165