snapd: Remove /snap symlink
The reason we needed /snap -> /var/lib/snapd/snap was due to OSTree
when the element was originally made. But we do not need it with
sysupdate.  So we can move to mounting into /snap directly. And we do
not need to patch anymore.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2632>