Created a new test suite for Mobile tests
The tests in the mobile test suite need to pass on resolutions sutable for mobile. For now the resolution is set to 480X800. I also added the mobile test suite in the main.pm file

modified the mobile_welcome file to set the scaling to 2x

Added a global variable to the testapi object called form_factor

This is used to set the state from the main.pm file for the testsuite which is currently running(mobile or desktop). The gnome_welcome test uses it to access the corresponding needles (mobile or desktop)

Created a new test suite for Mobile tests

The tests in the mobile test suite need to pass on resolutions sutable for mobile. For now the resolution is set to 480X800. I also added the mobile test suite in the main.pm file

modified the mobile_welcome file to set the scaling to 2x

Added a global variable to the testapi object called form_factor

This is used to set the state from the main.pm file for the testsuite which is currently running(mobile or desktop). The gnome_welcome test uses it to access the corresponding needles (mobile or desktop)

Modified 19 core app tests to test their mobile versions.

I did not create separate tests for the mobile versions because the tests are basically the same. The difference is in the needles. So I created separate needles for mobile versions of the apps

Made some progress on adding a monitors.xml file in order to control the scaling

Added another console and made changes to the fractional scaling

the console is called gdm and it points to tyy-console 1

Created a function called resize_app_to_mobile

This function uses looking glass to resize the window of the focused app. I then call the function in individual tests, checking if the testsuite is mobile before calling the function on the apps before testing them