vm-secure: Add service to set chassis type on first boot
This is useful to test on VMs. Systemd sets tyhe chassis type as "vm".
But when testing we might want to set it to "desktop", "handset" or
"tablet". To do that we can now add a kernel parameter, for example
`systemd.wants=gnomeos.set-chassis-type@handset.service`. This kernel
parameter can be injecting through type 11 smbios.