tracker-miners: Disable Landlock
This is a security restriction that is new in the Tracker Miners 3.7
series. It is preventing tracker-extract-3 from working in GNOME OS,
see: <https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/310>

This is blocking some work on testing search, and I don't currently
know how to diagnose the problem further so let's work around it.