Override noopenh264
We accidentally failed to update to OpenH264 2.3.1. Although 2.4 has
the same ABI as 2.3, it's a different ABI from 2.2. This means we need
to rebuild everything that uses OpenH264 in order to upgrade.

Accordingly, this change cannot be backported to stable runtimes.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2723>