Build GNOME OS with x86_64_v3
We need to keep the flatpak runtime in v1 sadly for now,
as it's expected to run anywhere.

However, we can bump the baseline requirment of GNOME OS
to x86_64_v3 since we already depend on UEFI and soon
tpm2 hardware which is from a similar hardware era.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2794>