gnome_welcome: Work around unreliable keynav in gnome-initial-setup
The issue with timezone entry occasionally failing appears to be due
to widget focus being random. If the focus is already in the
search_entry box then pressing 'tab' moves it *out* and then type_string
doesn't enter text there.

Really the widget focus should be consistent, but that is not simple
to achieve: see
<https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/156>.

Work around by clicking on the search_entry before typing.

Fixes: https://gitlab.gnome.org/GNOME/openqa-tests/-/issues/31