Build gnome-keyring against libsystemd
With this change, systemd user units gnome-keyring-daemon.service and
gnome-keyring-daemon.socket will be installed, and the daemon itself
will support systemd socket activation.

The knock-on effect is that pam_systemd will no longer spawn the daemon
directly when it is not running – instead, it will rely on systemd to do
that. This allows gnome-initial-setup-copy-worker.service to be
explicitly sequenced Before=gnome-keyring-daemon.service, avoiding a
race between the copy-worker copying the new user's keyring from
/run/gnome-initial-setup, and gnome-keyring-daemon starting up.

Pass --with-systemd explicitly to avoid silently falling back to
building without systemd support, as in the past.

This is the configuration used by every other distribution I checked:
Fedora, Debian, Arch, Ubuntu.

https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/249
https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/250
https://gitlab.gnome.org/GNOME/gnome-keyring/-/issues/137

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2839>