vm-secure/signed-boot.bst: depend strictly on usr-image
signed-boot.bst is a "manual" element rather than a "script" element,
this allows it to have sources, but it also make it not rebuild strictly
(as it assumes the contents of its artifacts are the built sources).

However, signed-boot.bst copies files from usr-image.bst, so it should
mark the dependency as strict so that it gets rebuilt whenever
usr-image.bst changes.

Fixes https://gitlab.gnome.org/GNOME/gnome-build-meta/-/issues/828

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2867>