systemd: Remove dependency for bpf
We now also need bpftool. The current build of libbpf was taken from a
git repository that is not in sync we our version of the kernel and is
weird to build.  We should add libbpf and bpftool in freedesktop-sdk
instead, and build them from the source code the kernel in use.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2742>