50-gnome.preset: Update systemd services
This update to the preset file for systemd includes the following
changes:
* Minor cleanup of English and neatness (self-explanatory)
* Disabling the systemd-firstboot.service
* Disabling the systemd-homed-firstboot.service

The systemd-firstboot.service was already disabled (see fdfbf177).
The difference is that we've moved it into the preset file. In the
next commit, we will reverse the actual commit changes.

The systemd-homed-firstboot.service has been additionally disabled
because it now integrates additional logic that requires more than just
the presence of a root user. If there are no regular user accounts
existing, the service prompts interactively for the creation of one,
complicating automated setups.