incubator: Add Decibels
Decibels is a new audio player that's currently in the incubation
stage of our app lifecycle. Since it's aiming to become a core app,
we want to be able to build it alongside our other core apps.

This commit adds a new stack for Incubator apps, ensures
it's built in the VM, and adds Decibels to this set.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3022>