utils/test_phone_image.py: Remove redundant module check
Testing for LEDS_CLASS_FLASH is redundant because the config was already
part of the original freedesktop-sdk kernel before we integrated the
kernel inside !2455.

The kernel config tests are supposed to represent the additional configs
we inserted into the kernel during integration. Hence, the test for
LEDS_CLASS_FLASH is, by definition, redundant.

Please view the following link for more information:
* https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/merge_requests/20683