core-deps/udisks2: Add patch for NVMe drive size
udisks2 has an issue with detecting NVMe drives. This
causes an issue in GNOME OS where an available disk
shows a size of 0 bytes, preventing installation.

This commit applies a patch from an upstream PR that solves
the issue. I've tested this patch with the installer, and
all seems to work well - including installation and
apps like Disks and Baobab.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3026>