ci: Split openQA tests into two jobs
We now have test-openqa-desktop and test-openqa-installer. These
will run in parallel on every gnome-build-meta pipeline. The
upload of the "latest" GNOME OS image will happen if
test-openqa-installer passes, the result of test-openqa-desktop is
ignored.

For various reasons[1] the full openQA test suite does not pass reliably.
The motivation here is to have a more minimal go/no-go test: can the
.iso install a GNOME OS build which successfully launches gnome-shell?

The remaining tests, which ensure the correct font size in each of the
core apps and so on, will no longer block the "tested" image from
uploading.

The slowest part of the openQA testsuite is usually downloading test
media[2].  The jobs are split so that test-openqa-installer only
requires the installer.iso and test-openqa-desktop only requires the
disk.img. So the two new jobs will complete faster than the old
test-s3-image job that downloaded both disk.img and installer.iso.

1. https://gitlab.gnome.org/GNOME/openqa-tests/-/issues/62
2. https://gitlab.gnome.org/GNOME/openqa-tests/-/issues/120

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3143>