utils/update-refs: Automatically skip problematic ref updates
Add a list of elements to ignore during tracking. This is only
for elements that we know can't be updated atm due to an issue
or other changes pending. Keep the list in the script itself
so we don't have to keep track of them individually.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3268>