ci: Don't override ARCH variable for x86_64
In f8153cede3b65900d64aab9c196fc7f5bc8361a4 we changed
the ARCH variable to be X86_64_V3 so we can simplify
the build_elements scripts by matching on that.

However we are also using the ARCH variable currently
to determine the filename of the disk images we upload.

Instead of using arch, introduce a new X86_64_V3 and
check that to determine what targets to build.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3332>