oci/*.bst: disable gzip compression
Compression is the thing taking most of the time, and isn't used later
on because we pass it to podman (which will decompress it anyway).

The only benefit we have for compression is storage space on the CAS,
but we can disable pushing the images to CAS if necessary.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/3453>