live-image: Add pret-a-booter installation from live system
This is a demo for installing the live system onto disk and not even
needing to reboot.

We boot the live system with a root btrfs in ramdisk. Then if we
decide we like the live system, we repart a disk with encryption, then
we move the live root disk into it. It is quick, and the user does not
even have to reboot.

To test it, run in the live system (if disk to install is /dev/vda):
run0 /usr/libexec/gnomeos/pret-a-booter-install /dev/vda