live-image: Add smooth installation from live system
This is a demo for installing the live system onto disk and not even
needing to reboot.

We boot the live system with a root btrfs in ramdisk. Then if we
decide we like the live system, we repart a disk with encryption, then
we move the live root disk into it. It is quick, and the user does not
even have to reboot.

To test it, when using a cdrom use:
/usr/libexec/gnomeos/do-the-magic-install /dev/loop0 /dev/vda
When using 2 virtio disks (first being the live):
/usr/libexec/gnomeos/do-the-magic-install /dev/vda /dev/vdb